fix: table 분리 완료
