+++ .gitignore
@@ -12,4 +12,4 @@ db.sqlite3-journal
 media
 venv
 .idea
-migrations/
+migrations
