+++ .idea/workspace.xml
@@ -5,8 +5,20 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="a274433b-f04f-4270-8662-5c40252cefdb" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/backend/django_back/capsules/views.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/capsules/views.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/capsules/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/capsules/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/django_back/settings.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/django_back/settings.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/admin.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/admin.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/apps.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/apps.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/tests.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/tests.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/music/views.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/musics/views.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/stories/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/stories/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/themes/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/themes/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/backend/django_back/videos/models.py" beforeDir="false" afterPath="$PROJECT_DIR$/backend/django_back/videos/models.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/docker-compose.yml" beforeDir="false" afterPath="$PROJECT_DIR$/docker-compose.yml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
