+++ backend/django_back/capsules/models.py
@@ -5,16 +5,18 @@ from django.db import models
 
 from baseapp.models import BaseModel
 from users.models import User
-from videos.models import Video
-from stories.models import Story
+from themes.models import Theme
+# from videos.models import Video
+# from stories.models import Story
 
 class Capsule(BaseModel):
     capsule_id = models.AutoField(primary_key=True)
-    # user_id = models.ManyToManyField(User, db_column='user_id', through='UserCapsule')
+    # user_id = models.ForeignKey(User, on_delete=models.CASCADE)
     # test를 위해 relation 임시 제거
-    user_id = models.IntegerField(null=True)
-    # theme_id = models.ForeignKey('Theme', on_delete=models.CASCADE, db_column='user_id')
-    theme_id = models.IntegerField(null=True)
+    # user_id = models.IntegerField(null=True)
+    creator_id = models.ForeignKey(User, on_delete=models.CASCADE)
+    theme_id = models.ForeignKey(Theme, on_delete=models.CASCADE)
+    # theme_id = models.IntegerField(null=True)
     capsule_name = models.CharField(max_length=60)
     due_date = models.DateTimeField()
     limit_count = models.IntegerField(null=True, validators=[MinValueValidator(1), MaxValueValidator(30)])
@@ -22,17 +24,6 @@ class Capsule(BaseModel):
 
     class Meta:
         db_table = 'capsule'
-
-
-# class Theme(BaseModel):
-#     theme_id = models.AutoField(primary_key=True)
-#     theme_name = models.CharField(max_length=60)
-#     theme_img_url = models.CharField(max_length=255)
-#
-#     class Meta:
-#         db_table = 'theme'
-
-
 class UserCapsule(models.Model):
     user_capsule_id = models.AutoField(primary_key=True)
     capsule_id = models.ForeignKey(Capsule, on_delete=models.CASCADE)
@@ -42,6 +33,17 @@ class UserCapsule(models.Model):
         db_table = 'user_capsule'
 
 
+# class Theme(BaseModel):
+#     theme_id = models.AutoField(primary_key=True)
+#     theme_name = models.CharField(max_length=60)
+#     theme_img_url = models.CharField(max_length=255)
+#
+#     class Meta:
+#         db_table = 'theme'
+#
+#
+#
+#
 # class Story(BaseModel):
 #     story_id = models.AutoField(primary_key=True)
 #     capsule_id = models.ForeignKey('Capsule', on_delete=models.CASCADE)
@@ -53,7 +55,7 @@ class UserCapsule(models.Model):
 #
 #     class Meta:
 #         db_table = 'story'
-
+#
 # class Video(BaseModel):
 #     video_id = models.AutoField(primary_key=True)
 #     capsule_id = models.ForeignKey(Capsule, on_delete=models.CASCADE)
@@ -62,16 +64,16 @@ class UserCapsule(models.Model):
 #     story_video_url = models.CharField(max_length=255)
 #
 #     class Meta:
-
-
-class StoryVideo(models.Model):
-    story_video_id = models.AutoField(primary_key=True)
-    story_id = models.ForeignKey(Story, on_delete=models.CASCADE)
-    video_id = models.ForeignKey(Video, on_delete=models.CASCADE)
-
-    class Meta:
-        db_table = 'story_video'
-
+#
+#
+# class StoryVideo(models.Model):
+#     story_video_id = models.AutoField(primary_key=True)
+#     story_id = models.ForeignKey(Story, on_delete=models.CASCADE)
+#     video_id = models.ForeignKey(Video, on_delete=models.CASCADE)
+#
+#     class Meta:
+#         db_table = 'story_video'
+#
 # class Music(BaseModel):
 #     music_id = models.AutoField(primary_key=True)
 #     music_name = models.CharField(max_length=60)
