+++ backend/django_back/musics/apps.py
@@ -3,6 +3,6 @@ from django.apps import AppConfig
 
 class MusicConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'music'
+    name = 'musics'
