+++ backend/django_back/themes/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from baseapp.models import BaseModel
-# Create your models here.
 class Theme(BaseModel):
     theme_id = models.AutoField(primary_key=True)
     theme_name = models.CharField(max_length=60)
